Pull quantile stuff out of data_access because the quantile stuff is porky enough to stand on its own and data_access is a port of a Java test file anyway.
